commiting

+++ mems.h
@@ -181,7 +181,7 @@ Parameter: MeMS Virtual address (that is created by MeMS)
 Returns: MeMS physical address mapped to the passed ptr (MeMS virtual address).
 */
 void *mems_get(void*v_ptr){
-    int a=(int)v_ptr;
+    int a=(size_t)v_ptr;
     Chain* CurrentChain=free_list_head;
     while(CurrentChain!=NULL){
         if(a<=CurrentChain->offset+CurrentChain->size){
